Add SIGCHLD to clone flags
Presumably this fixes tailhook/vagga#119

We don't need SIGCHLD here, but we must handle it anyway because on most
systems the signal deliver status is reset after execve. But it seems
that on some systems, as mentioned in the issue above, signal delivery
status may not be reset.

Also if we try to wait on child which doesn't have SIGCHLD flag zombie
gets killed before we are able to `wait()` for it. This results in
ECHILD error. So `Child::wait()` method fails for no good reason.